refactor(riot): migrate to TS
