+++ app/riot/package.json
@@ -53,10 +53,11 @@
     "ts-dedent": "^2.0.0"
   },
   "devDependencies": {
-    "@babel/plugin-transform-modules-commonjs": "^7.12.1",
-    "@babel/preset-env": "^7.12.1",
-    "@babel/preset-flow": "^7.12.1",
-    "@babel/preset-react": "^7.12.1"
+    "@types/riot": "^3.6.2",
+    "riot": "^3.13.2",
+    "riot-compiler": "^3.6.0",
+    "riot-hot-reload": "^1.0.0",
+    "riot-tag-loader": "^2.1.0"
   },
   "peerDependencies": {
     "@babel/core": "*",
