+++ app/riot/src/client/preview/rendering/compiledButUnmounted.ts
@@ -1,5 +1,5 @@
 import { mount } from 'riot';
 
-export default function renderCompiledButUnmounted(component) {
+export default function renderCompiledButUnmounted(component: any) {
   mount('root', component.tagName, component.opts || {});
 }
