+++ app/riot/src/client/preview/rendering/raw.ts
@@ -2,7 +2,7 @@ import { mount, tag2 as tag } from 'riot';
 import compiler from 'riot-compiler';
 import { alreadyCompiledMarker, getRidOfRiotNoise } from '../compileStageFunctions';
 
-export default function renderRaw(sourceCode) {
+export default function renderRaw(sourceCode: string) {
   const tag2 = tag;
   // eslint-disable-next-line no-eval
   eval(
