+++ app/riot/src/typings.d.ts
@@ -1,0 +1,8 @@
+declare module '@storybook/core/*';
+declare module 'global';
+declare module 'riot-compiler';
+declare module 'riot' {
+  const tag2: Function;
+  const mount: Function;
+  const unregister: Function;
+}
